refactor: isolate renders to reduce requests
