+++ src/components/SelectHotels/ValidateHotels.js
@@ -1,10 +1,7 @@
-import RoomsOfHotel from '../Rooms';
-import styled from 'styled-components';
 import { useEffect, useState } from 'react';
-import useHotels from '../../hooks/api/useHotels';
-import HotelDescription from './HotelDescription';
 import useTicket from '../../hooks/api/useTicket';
 import ErrorWrapper from '../ErrorWrapper';
+import HotelsRender from './Hotels';
 
 export const TicketStatus = Object.freeze({
   PAID: 'PAID',
@@ -17,9 +14,6 @@ export default function ValidateHotel() {
     isAllowed: true,
     message: '',
   });
-
-  const [content, setContent] = useState(<></>);
-  const [selectHotel, setSelectHotel] = useState(null);
 
   useEffect(() => {
     if (ticket) verifyPermission();
@@ -47,38 +41,5 @@ export default function ValidateHotel() {
     }
   }
 
-  const hotels = useHotels();
-
-  useEffect(() => {
-    if (hotels.hotels) {
-      setContent(
-        <>
-          <Text>Primeiro, escolha seu hotel</Text>
-          <ChooseHotels>
-            {hotels.hotels.map((value, index) => (
-              <HotelDescription value={value} key={index} selectHotel={selectHotel} setSelectHotel={setSelectHotel} />
-            ))}
-          </ChooseHotels>
-          {!selectHotel ? '' : <RoomsOfHotel hotelId={selectHotel} />}
-        </>
-      );
-    }
-  }, [hotels.hotelsLoading, selectHotel]);
-
-  return <>{!authStatus.isAllowed ? <ErrorWrapper>{authStatus.message}</ErrorWrapper> : content}</>;
+  return <>{!authStatus.isAllowed ? <ErrorWrapper>{authStatus.message}</ErrorWrapper> : <HotelsRender />}</>;
 }
-
-const Text = styled.div`
-  font-family: 'Roboto', sans-serif;
-  font-style: normal;
-  font-weight: 400;
-  font-size: 20px;
-  line-height: 23px;
-  color: #8e8e8e;
-  margin: 0 0 10px 0;
-`;
-
-const ChooseHotels = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-`;
